requirements/base.txt
arrow==0.10.0
astroid==1.6.1
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
disqus==0.0.4
django-disqus==0.5
Django==1.11.7
django-debug-toolbar==1.9.1
django-emoticons==1.2
djangorestframework==3.6.3
django-forms-bootstrap==3.1.0
django-imagekit==4.0.2
django-material==1.2.5
django-registration==2.2
django-tinymce==2.6.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
Pillow==4.1.1
pylint==1.8.2
python-dateutil==2.6.0
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.4
wheel==0.24.0
django-paypal==0.2.7
stripe==1.77.2
requests==2.18.4
urllib3==1.22
wrapt==1.10.11
